pyproject.toml
[project]
name = "pidforeducation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.3",
    "pillow>=11.1.0",
    "pygame>=2.6.1",
]
